changed f64 to f32

+++ src/geometry/point.rs
@@ -7,13 +7,13 @@ use std::ops::{Add, Sub, Mul, Div};
 /// A `Point` represents a position in space
 #[derive(Clone, Default, Copy)]
 pub struct Point {
-    pub x: f64,
-    pub y: f64
+    pub x: f32,
+    pub y: f32
 }
 
 impl Point {
     /// Returns a new `Point` with the given coordinates
-    pub fn new(x: f64, y: f64) -> Point {
+    pub fn new(x: f32, y: f32) -> Point {
         Point { x: x, y: y }
     }
 
@@ -26,13 +26,13 @@ impl Point {
     }
 
     /// Returns the squared distance from this point to the given one
-    pub fn squared_distance_to(&self, target: &Point) -> f64 {
+    pub fn squared_distance_to(&self, target: &Point) -> f32 {
         (self.x - target.x) * (self.x - target.x)
             + (self.y - target.y) * (self.y - target.y)
     }
 
     /// Rotates the point through the origin in the given angle (radians)
-    pub fn rotate(mut self, radians: f64) -> Point {
+    pub fn rotate(mut self, radians: f32) -> Point {
         let radius = (self.x * self.x + self.y * self.y).sqrt();
         let point_angle = (self.y / self.x).atan();
         let final_angle = point_angle + radians;
@@ -49,7 +49,7 @@ impl Point {
     }
 
     /// Checks if this point is contained in a circle
-    pub fn intersect_circle(self, center: &Point, radius: f64) -> bool {
+    pub fn intersect_circle(self, center: &Point, radius: f32) -> bool {
         (self.x - center.x).powi(2) +
             (self.y - center.y).powi(2) < radius.powi(2)
     }
@@ -74,11 +74,11 @@ impl Add for Point {
     }
 }
 
-/// Implements the '+' operator for Point + f64
-impl Add<f64> for Point {
+/// Implements the '+' operator for Point + f32
+impl Add<f32> for Point {
     type Output = Point;
 
-    fn add(self, _rhs: f64) -> Point {
+    fn add(self, _rhs: f32) -> Point {
         Point {
             x: self.x + _rhs,
             y: self.y + _rhs,
@@ -98,11 +98,11 @@ impl Sub for Point {
     }
 }
 
-/// Implements the '-' operator for Point - f64
-impl Sub<f64> for Point {
+/// Implements the '-' operator for Point - f32
+impl Sub<f32> for Point {
     type Output = Point;
 
-    fn sub(self, _rhs: f64) -> Point {
+    fn sub(self, _rhs: f32) -> Point {
         Point {
             x: self.x - _rhs,
             y: self.y - _rhs,
@@ -122,11 +122,11 @@ impl Mul for Point {
     }
 }
 
-/// Implements the '*' operator for Point * f64
-impl Mul<f64> for Point {
+/// Implements the '*' operator for Point * f32
+impl Mul<f32> for Point {
     type Output = Point;
 
-    fn mul(self, _rhs: f64) -> Point {
+    fn mul(self, _rhs: f32) -> Point {
         Point {
             x: self.x * _rhs,
             y: self.x * _rhs,
@@ -139,8 +139,8 @@ impl Div for Point {
     type Output = Point;
 
     fn div(self, _rhs: Point) -> Point {
-        assert!(_rhs.x != 0f64);
-        assert!(_rhs.y != 0f64);
+        assert!(_rhs.x != 0f32);
+        assert!(_rhs.y != 0f32);
         Point {
             x: self.x / _rhs.x,
             y: self.y / _rhs.y,
@@ -148,12 +148,12 @@ impl Div for Point {
     }
 }
 
-/// Implements the '/' operator for Point / f64:
-impl Div<f64> for Point {
+/// Implements the '/' operator for Point / f32:
+impl Div<f32> for Point {
     type Output = Point;
 
-    fn div(self, _rhs: f64) -> Point {
-        assert!(_rhs != 0f64);
+    fn div(self, _rhs: f32) -> Point {
+        assert!(_rhs != 0f32);
         Point {
             x: self.x / _rhs,
             y: self.y / _rhs,
